requirements.txt
alpaca-py>=0.7.0
yfinance
